refactor

+++ NHF/Source/Utilities/Math/Transitions.cpp
@@ -1,7 +1,5 @@
 #include "Transitions.hpp"
 
-#include "Transitionable.hpp"
-
 
 void Transition::update() {
 	if (_isActive && !_isPaused) {
@@ -24,14 +22,14 @@ namespace Transitions {
 			sf::Vector2f progression;
 			if (elapsedTime <= getDurationTime() / 2) {
 				progression = {
-					_accX / 2.f * math::square(elapsedTime) - _accX / 2.f * math::square(getElapsedTime()),
-					_accY / 2.f * math::square(elapsedTime) - _accY / 2.f * math::square(getElapsedTime())
+					_accX / 2.f * static_cast<float>(math::square(elapsedTime)) - _accX / 2.f * static_cast<float>(math::square(getElapsedTime())),
+					_accY / 2.f * static_cast<float>(math::square(elapsedTime)) - _accY / 2.f * static_cast<float>(math::square(getElapsedTime()))
 				};
 			}
 			else {
 				progression = {
-					_accX / 2.f * math::square(getDurationTime() - getElapsedTime()) - _accX / 2.f * math::square(getDurationTime() - elapsedTime),
-					_accY / 2.f * math::square(getDurationTime() - getElapsedTime()) - _accY / 2.f * math::square(getDurationTime() - elapsedTime)
+					_accX / 2.f * static_cast<float>(math::square(getDurationTime() - getElapsedTime())) - _accX / 2.f * static_cast<float>(math::square(getDurationTime() - elapsedTime)),
+					_accY / 2.f * static_cast<float>(math::square(getDurationTime() - getElapsedTime())) - _accY / 2.f * static_cast<float>(math::square(getDurationTime() - elapsedTime))
 				};
 			}
 			return progression;
@@ -52,7 +50,7 @@ namespace Transitions {
 
 		Jump::Jump(Transitionable* object) :
 			Transition{ object, [this](int elapsedTime) -> sf::Vector2f {
-				float progressionF = _acc / 2.f * math::square(elapsedTime) + _velocity * static_cast<float>(elapsedTime) - _lastProgress;
+				float progressionF = _acc / 2.f * static_cast<float>(math::square(elapsedTime)) + _velocity * static_cast<float>(elapsedTime) - _lastProgress;
 				sf::Vector2f progression = sf::Vector2f{ _direction } *progressionF;
 
 				_lastProgress += progressionF;
